pages and partners services renamed. App and env params added
